Rewrite secret file to warn users against copying
Problem: I keep seeing new SSB users who want same-as and decide to
experiment with copying their secret file to multiple devices at the
same time. This seems like it should work, and this sort of hacky ethos
would usually be encouraged, but it's risky and will inevitably fork the
feed.

Solution: Rewrite the secret file to move the warnings to the top,
adding a warning about copying the secret file to a second device.